mod: extract shared test helpers into test/helpers.ts
Deduplicate req, multipartReq, sseReq, assertErrorEnvelope, and
generateSilentWav into a shared helpers module used by both
serve.test.ts and e2e.test.ts via createTestClient(base).

Made-with: Cursor
